+++ packages/zowe-explorer/src/dataset/actions.ts
@@ -142,9 +142,10 @@ export async function allocateLike(
     // Refresh tree and open new node, if applicable
     if (!currSession) {
         currSession = datasetProvider.mSessionNodes.find(
-            (thisSession) => thisSession.label.toString() === profile.name
+            (thisSession) => thisSession.label.toString().trim() === profile.name
         );
     }
+
     const theFilter = await datasetProvider.createFilterString(newDSName, currSession);
     currSession.tooltip = currSession.pattern = theFilter.toUpperCase();
     datasetProvider.addSearchHistory(theFilter);
@@ -246,35 +247,24 @@ export async function deleteDatasetPrompt(
     selectedNodes = selectedNodes.filter((val) => !childArray.includes(val));
 
     if (includedSelection || !node) {
-        // Filter out sessions, favorite nodes, or information messages
+        // Filter out sessions and information messages
         nodes = selectedNodes.filter(
             (selectedNode) =>
                 selectedNode.getParent() &&
-                !contextually.isFavorite(selectedNode) &&
-                !contextually.isFavorite(selectedNode.getParent()) &&
                 !contextually.isSession(selectedNode) &&
                 !contextually.isInformation(selectedNode)
         );
     } else {
-        if (
-            node.getParent() &&
-            !contextually.isFavorite(node) &&
-            !contextually.isFavorite(node.getParent()) &&
-            !contextually.isSession(node) &&
-            !contextually.isInformation(node)
-        ) {
+        if (node.getParent() && !contextually.isSession(node) && !contextually.isInformation(node)) {
             nodes = [];
             nodes.push(node);
         }
     }
 
-    // Check that there are items to be deleted, this can be caused by trying to delete favorites right now
+    // Check that there are items to be deleted
     if (!nodes || nodes.length === 0) {
         vscode.window.showInformationMessage(
-            localize(
-                "deleteDatasetPrompt.nodesToDelete.empty",
-                "Deleting data sets and members from the Favorites section is currently not supported."
-            )
+            localize("deleteDatasetPrompt.nodesToDelete.empty", "No data sets selected for deletion, cancelling...")
         );
         return;
     }
@@ -425,6 +415,7 @@ export async function createMember(
         }
         parent.dirty = true;
         datasetProvider.refreshElement(parent);
+
         openPS(
             new ZoweDatasetNode(
                 name,
@@ -438,6 +429,13 @@ export async function createMember(
             true,
             datasetProvider
         );
+
+        // Refresh corresponding tree parent to reflect addition
+        const otherTreeParent = datasetProvider.findEquivalentNode(parent, contextually.isFavorite(parent));
+        if (otherTreeParent != null) {
+            datasetProvider.refreshElement(otherTreeParent);
+        }
+
         datasetProvider.refresh();
     }
 }
@@ -783,57 +781,55 @@ async function handleUserSelection(newDSProperties, dsType): Promise<string> {
  * Shows data set attributes in a new text editor
  *
  * @export
- * @param {IZoweDatasetTreeNode} parent - The parent Node
+ * @param {IZoweDatasetTreeNode} node   - The node to show attributes for
  * @param {DatasetTree} datasetProvider - the tree which contains the nodes
  */
-export async function showDSAttributes(
-    parent: api.IZoweDatasetTreeNode,
+export async function showAttributes(
+    node: api.IZoweDatasetTreeNode,
     datasetProvider: api.IZoweTree<api.IZoweDatasetTreeNode>
 ) {
-    await datasetProvider.checkCurrentProfile(parent);
+    await datasetProvider.checkCurrentProfile(node);
     if (Profiles.getInstance().validProfile !== api.ValidProfileEnum.INVALID) {
-        const label = parent.label as string;
-        globals.LOG.debug(localize("showDSAttributes.debug", "showing attributes of data set ") + label);
+        const label = node.label as string;
+        globals.LOG.debug(localize("showAttributes.debug", "showing attributes for ") + label);
         let attributes: any;
         try {
-            attributes = await ZoweExplorerApiRegister.getMvsApi(parent.getProfile()).dataSet(label, {
-                attributes: true,
-            });
-            attributes = attributes.apiResponse.items;
-            attributes = attributes.filter((dataSet) => {
-                return dataSet.dsname.toUpperCase() === label.toUpperCase();
-            });
-            if (attributes.length === 0) {
-                throw new Error(
-                    localize("showDSAttributes.lengthError", "No matching data set names found for query: ") + label
+            if (contextually.isDsMember(node)) {
+                const dsName = node.getParent().getLabel() as string;
+                attributes = await ZoweExplorerApiRegister.getMvsApi(node.getProfile()).allMembers(
+                    dsName.toUpperCase(),
+                    {
+                        attributes: true,
+                        pattern: label.toUpperCase(),
+                    }
                 );
+            } else {
+                attributes = await ZoweExplorerApiRegister.getMvsApi(node.getProfile()).dataSet(label, {
+                    attributes: true,
+                });
+            }
+            attributes = attributes.apiResponse.items;
+            if (contextually.isDs(node)) {
+                attributes = attributes.filter((dataSet) => {
+                    return dataSet.dsname.toUpperCase() === label.toUpperCase();
+                });
+            }
+            if (attributes.length === 0) {
+                throw new Error(localize("showAttributes.lengthError", "No matching names found for query: ") + label);
             }
         } catch (err) {
             globals.LOG.error(
-                localize("showDSAttributes.log.error", "Error encountered when listing attributes! ") +
+                localize("showAttributes.log.error", "Error encountered when listing attributes! ") +
                     JSON.stringify(err)
             );
             await errorHandling(
                 err,
-                parent.getProfileName(),
-                localize("showDSAttributes.error", "Unable to list attributes: ") + err.message
+                node.getProfileName(),
+                localize("showAttributes.error", "Unable to list attributes: ") + err.message
             );
             throw err;
         }
 
-        // shouldn't be possible for there to be two cataloged data sets with the same name,
-        // but just in case we'll display all of the results
-        // if there's only one result (which there should be), we will just pass in attributes[0]
-        // so that prettyJson doesn't display the attributes as an array with a hyphen character
-        const attributesText = zowe.imperative.TextUtils.prettyJson(
-            attributes.length > 1 ? attributes : attributes[0],
-            undefined,
-            false
-        );
-        // const attributesFilePath = path.join(ZOWETEMPFOLDER, label + ".yaml");
-        // fs.writeFileSync(attributesFilePath, attributesText);
-        // const document = await vscode.workspace.openTextDocument(attributesFilePath);
-        // await vscode.window.showTextDocument(document);
         const attributesMessage = localize("attributes.title", "Attributes");
         const webviewHTML = `<!DOCTYPE html>
         <html lang="en">
@@ -842,14 +838,26 @@ export async function showDSAttributes(
             <title>${label} "${attributesMessage}"</title>
         </head>
         <body>
-        ${attributesText.replace(/\n/g, "</br>")}
+        <table style="margin-top: 2em; border-spacing: 2em 0">
+        ${Object.keys(attributes[0]).reduce(
+            (html, key) =>
+                html.concat(`
+                <tr>
+                    <td align="left" style="color: yellow; font-weight: bold">${key}:</td>
+                    <td align="right" style="color: ${
+                        typeof attributes[0][key] === "number" ? "dodgerblue" : "white"
+                    }">${attributes[0][key]}</td>
+                </tr>
+        `),
+            ""
+        )}
+        </table>
         </body>
         </html>`;
-        const column = vscode.window.activeTextEditor ? vscode.window.activeTextEditor.viewColumn : undefined;
         const panel: vscode.WebviewPanel = vscode.window.createWebviewPanel(
             "zowe",
             label + " " + localize("attributes.title", "Attributes"),
-            column || 1,
+            vscode.window.activeTextEditor ? vscode.window.activeTextEditor.viewColumn : 1,
             {}
         );
         panel.webview.html = webviewHTML;
@@ -1001,18 +1009,20 @@ export async function submitMember(node: api.IZoweTreeNode) {
 export async function deleteDataset(node: api.IZoweTreeNode, datasetProvider: api.IZoweTree<api.IZoweDatasetTreeNode>) {
     let label = "";
     let fav = false;
+
+    const parent = node.getParent();
     try {
-        const parentContext = node.getParent().contextValue;
+        const parentContext = parent.contextValue;
         if (parentContext.includes(globals.FAV_SUFFIX)) {
             label = node.getLabel() as string;
             fav = true;
             if (parentContext.includes(globals.DS_PDS_CONTEXT + globals.FAV_SUFFIX)) {
-                label = node.getParent().getLabel().toString() + "(" + node.getLabel().toString() + ")";
+                label = parent.getLabel().toString() + "(" + node.getLabel().toString() + ")";
             }
         } else if (parentContext.includes(globals.DS_SESSION_CONTEXT)) {
             label = node.getLabel() as string;
         } else if (parentContext.includes(globals.DS_PDS_CONTEXT)) {
-            label = node.getParent().getLabel().toString() + "(" + node.getLabel().toString() + ")";
+            label = parent.getLabel().toString() + "(" + node.getLabel().toString() + ")";
         } else {
             throw Error(localize("deleteDataSet.invalidNode.error", "deleteDataSet() called from invalid node."));
         }
@@ -1046,11 +1056,23 @@ export async function deleteDataset(node: api.IZoweTreeNode, datasetProvider: ap
                 ses.dirty = true;
             }
         });
-        datasetProvider.removeFavorite(node);
     } else {
         node.getSessionNode().dirty = true;
-        datasetProvider.removeFavorite(node);
     }
+    datasetProvider.removeFavorite(node);
+
+    const isMember = contextually.isDsMember(node);
+
+    // If the node is a dataset member, go up a level in the node tree
+    // to find the relevant, matching node
+    const nodeOfInterest = isMember ? node.getParent() : node;
+    const parentNode = datasetProvider.findEquivalentNode(nodeOfInterest, fav);
+
+    if (parentNode != null) {
+        // Refresh the correct node (parent of node to delete) to reflect changes
+        datasetProvider.refreshElement(isMember ? parentNode : parentNode.getParent());
+    }
+
     datasetProvider.refreshElement(node.getSessionNode());
 
     // remove local copy of file
@@ -1060,7 +1082,7 @@ export async function deleteDataset(node: api.IZoweTreeNode, datasetProvider: ap
             fs.unlinkSync(fileName);
         }
     } catch (err) {
-        // do nothing
+        globals.LOG.warn(err);
     }
 }
 
@@ -1247,6 +1269,7 @@ export async function hMigrateDataSet(node: ZoweDatasetNode) {
         try {
             return ZoweExplorerApiRegister.getMvsApi(node.getProfile()).hMigrateDataSet(dataSetName);
         } catch (err) {
+            globals.LOG.error(err);
             vscode.window.showErrorMessage(err.message);
             return;
         }
@@ -1274,6 +1297,7 @@ export async function hRecallDataSet(node: ZoweDatasetNode) {
         try {
             return ZoweExplorerApiRegister.getMvsApi(node.getProfile()).hRecallDataSet(dataSetName);
         } catch (err) {
+            globals.LOG.error(err);
             vscode.window.showErrorMessage(err.message);
             return;
         }
@@ -1344,6 +1368,7 @@ export async function pasteMember(
                     { dsn: dataSetName, member: memberName }
                 );
             } catch (err) {
+                globals.LOG.error(err);
                 vscode.window.showErrorMessage(err.message);
                 return;
             }
@@ -1513,6 +1538,7 @@ export async function saveFile(doc: vscode.TextDocument, datasetProvider: api.IZ
             vscode.window.showErrorMessage(uploadResponse.commandResponse);
         }
     } catch (err) {
+        globals.LOG.error(err);
         vscode.window.showErrorMessage(err.message);
     }
 }