+++ packages/zowe-explorer/src/extension.ts
@@ -42,6 +42,7 @@ import { handleSaving } from "./utils/workspace";
 import { ZoweDatasetNode } from "./dataset/ZoweDatasetNode";
 import * as contextuals from "../src/shared/context";
 import { Job } from "./job/ZoweJobNode";
+import { getSelectedNodeList } from "./shared/utils";
 
 // Set up localization
 nls.config({
@@ -77,6 +78,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         globals.initLogger(context);
         globals.LOG.debug(localize("initialize.log.debug", "Initialized logger from VSCode extension"));
     } catch (err) {
+        globals.LOG.error(err);
         const errorMessage = localize(
             "initialize.log.error",
             "Error encountered while activating and initializing logger! "
@@ -88,6 +90,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         await initializeZoweFolder();
         await readConfigFromDisk();
     } catch (err) {
+        globals.LOG.error(err);
         const errorMessage = localize("initialize.profiles.error", "Error reading or initializing Zowe CLI profiles.");
         vscode.window.showWarningMessage(`${errorMessage}: ${err.message}`);
     }
@@ -442,7 +445,16 @@ function initDatasetProvider(context: vscode.ExtensionContext) {
         })
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.removeSession", async (node) => datasetProvider.deleteSession(node))
+        vscode.commands.registerCommand(
+            "zowe.ds.removeSession",
+            async (node: IZoweDatasetTreeNode, nodeList: IZoweDatasetTreeNode[]) => {
+                let selectedNodes = getSelectedNodeList(node, nodeList);
+                selectedNodes = selectedNodes.filter((sNode) => contextuals.isDsSession(sNode));
+                for (const select of selectedNodes) {
+                    datasetProvider.deleteSession(select);
+                }
+            }
+        )
     );
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.ds.removeFavorite", async (node, nodeList) => {
@@ -472,13 +484,12 @@ function initDatasetProvider(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand("zowe.ds.submitMember", async (node) => dsActions.submitMember(node))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.showDSAttributes", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isPdsNotFav(element)
+        vscode.commands.registerCommand("zowe.ds.showAttributes", async (node, nodeList) => {
+            const selectedNodes = getSelectedNodeList(node, nodeList).filter(
+                (element) => contextuals.isDs(element) || contextuals.isPds(element) || contextuals.isDsMember(element)
             );
             for (const item of selectedNodes) {
-                await dsActions.showDSAttributes(item, datasetProvider);
+                await dsActions.showAttributes(item, datasetProvider);
             }
         })
     );
@@ -619,9 +630,13 @@ function initUSSProvider(context: vscode.ExtensionContext) {
         )
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeSession", async (node: IZoweUSSTreeNode) =>
-            ussFileProvider.deleteSession(node)
-        )
+        vscode.commands.registerCommand("zowe.uss.removeSession", async (node: IZoweUSSTreeNode, nodeList) => {
+            let selectedNodes = getSelectedNodeList(node, nodeList);
+            selectedNodes = selectedNodes.filter((element) => contextuals.isUssSession(element));
+            for (const item of selectedNodes) {
+                ussFileProvider.deleteSession(item);
+            }
+        })
     );
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.uss.createFile", async (node: IZoweUSSTreeNode) =>
@@ -732,9 +747,9 @@ function initJobsProvider(context: vscode.ExtensionContext) {
         )
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.deleteJob", async (job, jobs) =>
-            jobActions.deleteCommand(jobsProvider, job, jobs)
-        )
+        vscode.commands.registerCommand("zowe.jobs.deleteJob", async (job, jobs) => {
+            await jobActions.deleteCommand(jobsProvider, job, jobs);
+        })
     );
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.jobs.runModifyCommand", (job) => jobActions.modifyCommand(job))
@@ -778,7 +793,13 @@ function initJobsProvider(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand("zowe.jobs.setPrefix", (job) => jobActions.setPrefix(job, jobsProvider))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.removeJobsSession", (job) => jobsProvider.deleteSession(job))
+        vscode.commands.registerCommand("zowe.jobs.removeJobsSession", (job, jobList) => {
+            let selectedNodes = getSelectedNodeList(job, jobList);
+            selectedNodes = selectedNodes.filter((element) => contextuals.isJobsSession(element));
+            for (const item of selectedNodes) {
+                jobsProvider.deleteSession(item);
+            }
+        })
     );
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.jobs.downloadSpool", async (node, nodeList) => {
@@ -877,16 +898,6 @@ function initSubscribers(context: vscode.ExtensionContext, theProvider: IZoweTre
             await theProvider.flipState(e.element, true);
         });
     }
-}
-
-function getSelectedNodeList(node: IZoweTreeNode, nodeList: IZoweTreeNode[]): IZoweTreeNode[] {
-    let resultNodeList: IZoweTreeNode[] = [];
-    if (!nodeList) {
-        resultNodeList.push(node);
-    } else {
-        resultNodeList = nodeList;
-    }
-    return resultNodeList;
 }
 
 /**