+++ packages/zowe-explorer/src/job/actions.ts
@@ -234,6 +234,7 @@ export async function modifyCommand(job: Job) {
         }
     } catch (error) {
         if (error.toString().includes("non-existing")) {
+            globals.LOG.error(error);
             vscode.window.showErrorMessage(
                 localize("jobActions.modifyCommand.apiNonExisting", "Not implemented yet for profile of type: ") +
                     job.getProfile().type
@@ -262,6 +263,7 @@ export async function stopCommand(job: Job) {
         }
     } catch (error) {
         if (error.toString().includes("non-existing")) {
+            globals.LOG.error(error);
             vscode.window.showErrorMessage(
                 localize("jobActions.stopCommand.apiNonExisting", "Not implemented yet for profile of type: ") +
                     job.getProfile().type
@@ -347,14 +349,15 @@ async function deleteSingleJob(job: IZoweJobTreeNode, jobsProvider: IZoweTree<IZ
         );
         return;
     }
+
     try {
         await jobsProvider.delete(job);
     } catch (error) {
         await errorHandling(error.toString(), job.getProfile().name, error.message.toString());
         return;
     }
-    await refreshAllJobs(jobsProvider);
-    vscode.window.showInformationMessage(localize("deleteCommand.job", "Job {0} deleted.", jobName));
+
+    vscode.window.showInformationMessage(localize("deleteCommand.job", "Job {0} was deleted.", jobName));
 }
 
 async function deleteMultipleJobs(
@@ -383,6 +386,7 @@ async function deleteMultipleJobs(
                 await jobsProvider.delete(job);
                 return job;
             } catch (error) {
+                globals.LOG.error(error);
                 return error;
             }
         })
@@ -396,7 +400,6 @@ async function deleteMultipleJobs(
         })
         .filter((result) => result !== undefined);
     if (deletedJobs.length) {
-        await refreshAllJobs(jobsProvider);
         vscode.window.showInformationMessage(
             localize(
                 "deleteCommand.multipleJobs",